Enhance authentication provider with onNotAuthenticated handler
-  Includes minor refactor in the IAnalyticalBackend & IAuthProvider location
   -  These two type defs joined into one file
   -  Backend capabilities split-off into a separate file
-  Expand IAuthenticationContext with 'backend'

JIRA: RAIL-2924
