+++ libs/sdk-backend-spi/src/backend/capabilities.ts
@@ -1,0 +1,74 @@
+// (C) 2019-2021 GoodData Corporation
+
+/**
+ * Analytical Backend communicates its capabilities via objects of this type. In return, the capabilities
+ * can then be used by applications to enable / disable particular features.
+ *
+ * @public
+ */
+export interface IBackendCapabilities {
+    /**
+     * Indicates whether the backend is capable to address objects using URIs
+     */
+    supportsObjectUris?: boolean;
+
+    /**
+     * Indicates whether the backend is capable to calculate and include totals in the resulting data view.
+     */
+    canCalculateTotals?: boolean;
+
+    /**
+     * Indicates whether the backend is capable to sort the result data view.
+     */
+    canSortData?: boolean;
+
+    /**
+     * Indicates whether the backend can recognize attribute elements by URI.
+     */
+    supportsElementUris?: boolean;
+
+    /**
+     * Indicates maximum result dimensions that the backend is able to produce.
+     */
+    maxDimensions?: number;
+
+    /**
+     * Indicates whether backend can export data to CSV file.
+     */
+    canExportCsv?: boolean;
+
+    /**
+     * Indicates whether backend can export data to Excel.
+     */
+    canExportXlsx?: boolean;
+
+    /**
+     * Indicates whether backend can transform an existing result into a different shape / sorting / totals.
+     */
+    canTransformExistingResult?: boolean;
+
+    /**
+     * Indicates whether backend can execute an existing, persistent insight by reference.
+     */
+    canExecuteByReference?: boolean;
+
+    /**
+     * Indicates whether backend supports adding CSV datasets and switching between them.
+     */
+    supportsCsvUploader?: boolean;
+
+    /**
+     * Indicates whether backend supports ranking filters.
+     */
+    supportsRankingFilter?: boolean;
+
+    /**
+     * Indicates whether backend supports element query parent filtering.
+     */
+    supportsElementsQueryParentFiltering?: boolean;
+
+    /**
+     * Catchall for additional capabilities
+     */
+    [key: string]: undefined | boolean | number | string;
+}
